[SIG-4531] Show notification when too many filters selected (#2249)
* add notification when too many filters selected

* add scrollIntoView for notification

* re-order folder structure

* implement PR feedback

* remove unncessecary SPDX

* fix tests

* implement PR feedback

* update typing
