+++ src/signals/incident-management/components/FilterForm/components/CheckboxGroup.tsx
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MPL-2.0
-// Copyright (C) 2020 - 2021 Gemeente Amsterdam
+// Copyright (C) 2020 - 2022 Gemeente Amsterdam
 import type { FC } from 'react'
 
 import Label from 'components/Label'
@@ -15,7 +15,7 @@ type CheckboxGroupProps = Partial<CheckboxListProps> & {
   name: string
 }
 
-const CheckboxGroup: FC<CheckboxGroupProps> = ({
+export const CheckboxGroup: FC<CheckboxGroupProps> = ({
   defaultValue,
   hasToggle,
   label,
@@ -51,5 +51,3 @@ CheckboxGroup.defaultProps = {
   defaultValue: [],
   hasToggle: true,
 }
-
-export default CheckboxGroup
