+++ src/signals/incident-management/components/FilterForm/index.ts
@@ -1,0 +1,3 @@
+import FilterForm from './FilterForm'
+
+export default FilterForm
