+++ src/signals/incident-management/components/FilterForm/styled.tsx
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MPL-2.0
-// Copyright (C) 2019 - 2021 Gemeente Amsterdam
+// Copyright (C) 2019 - 2022 Gemeente Amsterdam
 import FormFooter from 'components/FormFooter'
 
 import styled, { css } from 'styled-components'
@@ -86,4 +86,8 @@ export const FormFooterWrapper = styled(FormFooter)`
   button[type='button'] {
     order: 2;
   }
+`
+export const StyledNotification = styled.div`
+  margin-bottom: ${themeSpacing(5)};
+  scroll-margin-top: ${themeSpacing(5)};
 `
