+++ src/signals/incident-management/components/FilterForm/types.ts
@@ -1,0 +1,6 @@
+type FilterValue = string | string[]
+
+export type SelectedFilters = { [key: string]: FilterValue }
+export type StoredParams = {
+  [key: string]: FilterValue | number
+}
