+++ src/signals/incident-management/components/FilterForm/utils/constants.ts
@@ -1,0 +1,1 @@
+export const MAX_FILTER_LENGTH = 2700
