+++ src/signals/incident-management/components/FilterForm/utils/index.ts
@@ -1,0 +1,1 @@
+export { hasTooManyFiltersSelected } from './has-too-many-filters-selected'
